requirements.txt
beautifulsoup4==4.14.4
pytesseract==0.3.13
PyMuPDF==1.25.5
pdf2image==1.17.0
Pillow==11.2.1